feat(form): move Form class to src/form.ts, add type-safe only() method, export Form and HasValidate from index
