+++ index.ts
@@ -14,5 +14,6 @@ export { Component } from './src/component.js'
 export { EventBus } from './src/event_bus.js'
 export { Synth } from './src/synthesizers/synth.js'
 export { ViewComponent } from './src/view_component.js'
-export * from './src/decorators/index.js'
+export { Form } from './src/form.js'
 export { compose } from '@poppinss/utils'
+export { type HasValidate } from './src/features/support_validation/types.js'
