+++ package.json
@@ -18,7 +18,8 @@
     "./commands": "./build/commands/main.js",
     "./factories": "./build/src/factories/main.js",
     "./plugins/japa/api_client": "./build/src/plugins/japa/api_client.js",
-    "./plugins/japa/testing": "./build/src/plugins/japa/testing.js"
+    "./plugins/japa/testing": "./build/src/plugins/japa/testing.js",
+    "./decorators": "./build/src/decorators/index.js"
   },
   "scripts": {
     "test": "cross-env NODE_DEBUG=adonisjs:livewire c8 npm run quick:test",
